Add files via upload
Documented the Literature Review in the past tense, the word used in this section was 3074. The number of words that I chose to have for this section because I do not want to go over the limit. The referencing was included and some lines were taken and changed by myself to make a reference of what I found from the papers of other authors. First-person was not used in any part of this paper in fact I took extra care of not mentioning first-person words. Multiple figures and tables were included to robust my argument and knowledge that was gained from the research.