Add files via upload
Documented the methodology stage in the past tense, the words used for this chapter 1946, this amount was chosen because I do not want to go over the limit of 10,000 words for this whole paper. The main tests were included and described perfectly, the main details of the tests were provided, and even the presentation of the work. These were presented with an amount of information because I want the reader to know what is happening in every stage of the methodology chapter. No first-person were in use and tables and figures were presented to roust my argument and explain to the reader every single detail.