Add files via upload
Documented in the Past tense, used 3788 words, did not write 4000 because I did not want to go over the limit. Multiple tables, graphs, and screenshots were presented to robust my argument of the result gained from this research. Used some lines that were gathered from the papers as references to be able to compare and contrast between my results and the results of the papers I found. No passive style used in this section either and tables and figures were used to explain my result better