Add files via upload
Used past tense and present continuous for the part were future works is mentioned, in the introduction risks and reasons, why companies had any types of concerns for the migration of the cloud, were mentioned, the motivation behind the research was mentioned as well because security is ma passion, as the research objectives the questions were mentioned to kee the same objective throughout the entire paper, and a section in the introduction was left to describe the chapters of the whole paper. The conclusion has a summary of the results and findings, has a small summary for future works, also put some advantages and disadvantages of research. The Abstract had 248 words because again I do not want to go ever the word limit and multiple highlights from the dissertation were mentioned as well. References were included in some of the chapters, no first person was used in every chapter and tables and figures were not included because everything was showed in the previous chapters and explained well.